requirements.txt
openai>=1.6.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=7.4.0
streamlit>=1.28.0
